tests/submission/Infraware/Forms/contents/Forms/keygen_name.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>keygen_name</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
	 <p><keygen name="key" id='keygen_id'></keygen></p>
	</form>

	<script>
		
		var keygen = document.getElementById("keygen_id");
		var name = keygen.name;

		if (typeof(name) == "string") {
			test(function() {
				assert_equals(name, "key", "name attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("name attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
